refactor(model): 당첨 통계 클래스를 모델로 지정
